fixed flaky test SPringAmqpBatchIT - solved problem with bean creation order
